+++ lib/kitchen/directions/bake_preface/main.rb
@@ -3,8 +3,8 @@
 module Kitchen
   module Directions
     module BakePreface
-      def self.v1(book:)
-        V1.new.bake(book: book)
+      def self.v1(book:, title_element:)
+        V1.new.bake(book: book, title_element: title_element)
       end
     end
   end