+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add `BakeChapterSolutions` which bakes the free response solutions at the eoc (minor)
 * Changed locale `eoc_answer_key_title` to `answer_key_title` as it was only used in eob (major)
 * Added spec for translations (minor)
+* Remove summary attribute from numbered tables, add option to pass title element to `BakePreface` (minor)
 
 ## [5.0.0] - 2021-06-02